+++ cml-compiler/cml-language/src/main/java/cml/language/interfaces/tree/AfterRule.java
@@ -1,4 +1,4 @@
-package cml.language.interfaces.syntax_tree;
+package cml.language.interfaces.tree;
 
 @interface AfterRule
 {
