+++ cml-compiler/cml-language/src/main/java/cml/language/interfaces/tree/SyntaxTreeBuilder.java
@@ -1,4 +1,4 @@
-package cml.language.interfaces.syntax_tree;
+package cml.language.interfaces.tree;
 
 import cml.language.foundation.elements.SourceLocation;
